chore: upgrade `@aws-cdk/cloud-assembly-schema` to v41.2.0 (#34369)
### Issue # (if applicable)

N/A

### Reason for this change

This upgrade is supposed to be done as part of https://github.com/aws/aws-cdk/pull/34362. But the `@aws-cdk/cloud-assembly-schema` upgrade requires a manual update to the unit test so I am isolating this upgrade to this PR.

`v41.2.0` introduced a new line in the manifest file (PR: https://github.com/aws/aws-cdk-cli/pull/184), hence, the unit test was failing.

### Description of changes

Upgraded `@aws-cdk/cloud-assembly-schema` to `v41.2.0` and updated unit test.

### Describe any new or updated permissions being added

None

### Description of how you validated changes

Ran the build locally and no longer see the unit test failing

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*