feat(apigatewayv2): add dualstack support to domain name (#34111)
### Reason for this change

This builds on #34054. Apparently for a domain name to have IPv4 and IPv6
support, the `AWS::ApiGatewayV2::DomainName.DomainNameConfiguration` also
needs to set `IpAddressType`. This is distinct from the settings on the
`AWS::ApiGatewayV2::API` itself.



### Description of changes

This adds the [`ipAddressType` property][ipaddresstype] for `DomainNameConfiguration`.

[ipaddresstype]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-apigatewayv2-domainname-domainnameconfiguration.html#cfn-apigatewayv2-domainname-domainnameconfiguration-ipaddresstype



### Describe any new or updated permissions being added

N/A


### Description of how you validated changes

Unit & Integration tests have been added. These are the first tests for `DomainName`
for HTTP APIs and leverage the tooling for custom domains in integ tests.

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*