feat: support healthCheckGracePeriod in QueueProcessingFargateService (#34231)
Closes #34220

### Reason for this change

`QueueProcessingFargateService` supports the ECS deployment circuit breaker, but doesn't currently offer any way to configure the health check grace period used by the circuit breaker, as `FargateService` does. This PR rectifies that discrepancy.

### Description of changes

Added the `healthCheckGracePeriod` property to `QueueProcessingFargateService`, matching the corresponding property on `FargateService`.

### Describe any new or updated permissions being added

None.

### Description of how you validated changes

I've added a unit test to verify that the property is passed to the underlying `FargateService`.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*