feat(custom-resource): update default node runtime to node22 in custom resources for commercial, china and gov regions (#34319)
### Issue # (if applicable)

N/A

### Reason for this change

Node 20 is currently in maintenance phase and is slated to reach EOL on 2026-04-30 as indicated on https://nodejs.dev/en/about/releases/.

### Description of changes

Update `LATEST_NODE_RUNTIME_MAP` to `Node22` for commerical, china and US Gov regions. This will update CDK to use `nodejs22.x'` as the runtime for custom resources based on region. The `nodejs22.x` is not yet available in isolated regions, so it would be still be `Node18` for those regions.

### Describe any new or updated permissions being added

N/A

### Description of how you validated changes

Manually tested all individual custom resource handlers for create, update, and delete. 

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*