fix(route53): avoid setting unnecessary properties for Recovery Control type Route53 health check (#34272)
### Issue # (if applicable)

Closes #34262

### Reason for this change

Unnecessary properties are set for Route53 health checks for Recovery Control type.

### Description of changes

If HealthCheck type is Recovery Control, return `undefined` as default value for three properties (`FailureThreshold`, `RequestInterval`, and `MeasureLatency`).

### Describe any new or updated permissions being added

N/A

### Description of how you validated changes

Used `Match.absent()` to validate that the above three properties are not explicitly specified in existing unit tests. And added test cases to the integration tests

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*