fix: synthesis fails if `tree.json` exceeds 512MB (#34478)
In large applications with a lot of constructs, `tree.json` can grow to exceed 512MB, which means it can't be serialized anymore. 

In this PR, introduce the concept of a "subtree reference". Once a construct tree grows to exceed a fixed number of nodes we write subtrees to individual other files, and put a reference to those files into the original tree. 

The number of nodes can be configured with the context value `@aws-cdk/core.TreeMetadata:maxNodes`, and defaults to 500,000 (that assumes an average size of 1kB per node, which is an overestimate for safety. If we find that this number is too high in practice we may still lower it in the future).

Fixes https://github.com/aws/aws-cdk/issues/27261.

For unupdated consumers, there is graceful degradation here: the parts of the tree they will be able to see are cut off at certain tree depths, but only very large applications would be affected by this. Tree data consumers can be updated gradually to deal with these references.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*