fix(rds): fix incorrect import/export combination check in AuroraMysqlEngineVersion.of() (#33564)
Previously, AuroraMysqlEngineVersion.of() incorrectly determined whether import and export roles should be combined.

This caused inconsistent behavior when specifying engine versions manually using AuroraMysqlEngineVersion.of(...), leading to incorrect parameter settings in the DBClusterParameterGroup.

### Issue #33562

Closes #33562

### Reason for this change

The issue arises when specifying a custom Aurora MySQL version using AuroraMysqlEngineVersion.of(...).
The logic for determining whether import and export roles should be combined is incorrect.
It should be based on the major version.

This bug causes s3ExportBuckets to configure DBClusterParameterGroup incorrectly:
- It adds .aurora_select_into_s3_role instead of .aws_default_s3_role
- This behavior differs from using built-in versions like AuroraMysqlEngineVersion.VER_3_07_1

### Description of changes

If the major version is not "5.7", _combineImportAndExportRoles is set to true

### Describe any new or updated permissions being added

None.


### Description of how you validated changes

- Ran unit tests to confirm that the changes work as expected

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*