feat(events-targets): enable opt-in to use an IAM role for SnsTopic target (#34259)
### Issue # (if applicable)

There's currently no issue for this, but I'd be happy to create it.

### Reason for this change



Using an IAM role also fixes potential issues of failed published messages in case the target SNS topic is encrypted with KMS. According to this article, it's recommended to use execution roles for EventBridge targets. 

https://aws.amazon.com/about-aws/whats-new/2025/03/amazon-eventbridge-iam-execution-role-all-targets/

>  We recommend configuring execution roles for all your EventBridge targets to benefit from consistent permissions policies and (...)

### Description of changes

The `SnsTopic` target still has the same default behavior. However, it's now possible to opt-in to IAM role usage by setting the `authorizeUsingRole` property to `true`. It's also possible to provide a custom `role` as property.

### Describe any new or updated permissions being added

The `sns:Publish` action is allowed on the topicArn in case an IAM role is used.

### Description of how you validated changes

Unit tests were updated and added. 

There was already an integration test for the SnsTopic target without role usage.
I've added two more integration tests: one where `authorizeUsingRole` is set to true, and one where a custom role is provided.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*