fix(elasticloadbalancingv2): add validation on application listeners for certificates on HTTP protocol (#34233)
### Issue # (if applicable)


### Reason for this change

ElasticLoadBalancerV2 throw a 400 error if you try to append a certificate to a listener on port 80 (or protocol HTTP).
This PR brings this same validation to CDK

### Description of changes

Added a new check for the application protocol and the length of certificates, and if there is any certificate, throw a validation error.
Also, added a test for this case.

### Describe any new or updated permissions being added

### Description of how you validated changes

### Checklist

 My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license