chore: require cli integ test when changing/adding a feature flag. (#34584)
A feature flag has the potential to impact the behavior of all CDK apps, including those defined as part our [CLI integration test package](https://github.com/aws/aws-cdk-cli/tree/main/packages/%40aws-cdk-testing/cli-integ). 

For example: https://github.com/aws/aws-cdk/pull/34556

This PR made it so deploying CDK apps with a lambda function now also requires permissions to create log groups. This is not a breaking change for the CDK, but it did broke in our pipeline because of some permission limitations we impose on it. 

It could have been caught in the test pipeline.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*