feat(region-info): throw `ValidationError` instead of untyped Errors (#33384)
### Issue 

Relates to #32569 

### Description of changes

`ValidationErrors` where apprporiate

Note that files in the `build-tools` directory are excluded from the linter rules, since they are only used at build time.
I moved the `before` function to `build-tools` because it's not exported and only used at build time. Importing the `core` module to a build-time script doesn't work as it would add A LOT of additional files to the build-time hot path and that just  completely blows the build script up 💥 (aside from it being totally unnecessary anyway).

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

Existing tests. Exemptions granted as this is a refactor of existing code.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*