feat(eks): added support for `bootstrapSelfManagedAddons` (#33597)
### Issue # (if applicable)

Closes #30792

### Reason for this change

The feature enables support to create clusters without the default networking add-ons

### Description of changes



### Describe any new or updated permissions being added

Added the prop bootstrapSelfManagedAddons to the cluster and incremented the eks client version. Also validated that existing if bootstrapSelfManagedAddons is undefined to true or vice versa does not replace the cluster as the default is `true`.


### Description of how you validated changes

Validated the changes against an existing cluster and made sure it is not replaced unless the change is from true to false or vice versa. 

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*