fix(logs): exposed metric from the metric filter will now include the dimension map (#34648)
### Issue # (if applicable)

Closes #34643 

### Reason for this change

When the metric filter is created, you can specify the dimension map for the metric. When you extract the metric out of the metric filter, the extracted metric will not include the dimension map, and you will be forced to manually add it to the metric.

### Description of changes

This PR preserves the dimension map used when creating the metric filter and then applies it back when extracting the metric using the `MetricFilter.metric()` API. 

### Describe any new or updated permissions being added

N/A


### Description of how you validated changes

Added new unit and integration test that verify the extracted metric still includes the original dimension map used when creating the metric filter.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*