feat(codebuild): add custom instance type and VPC to Fleets
https://aws.amazon.com/about-aws/whats-new/2025/04/aws-codebuild-ec2-instance-type-configurable-storage-size/

CodeBuild now supports specifying specific EC2 instance types to serve
as fleet compute.

Add this support to the Fleet construct by way of adding
CUSTOM_INSTANCE_TYPE to the FleetComputeType enum, and `instanceType` to
`ComputeConfiguration`.

Also, add VPC support to Fleet. This mirrors the VPC support in Project.
When using Fleets, the VPC configured on the Project does nothing. Only
the VPC on the Fleet applies. This required adding a Role to the Fleet
to handle provisioning EC2 network interfaces in the configured VPC.