fix(pipelines): use display name for `Asset` actions when `publishAssetsInParallel` is false (#34049)
### Issue # (if applicable)

Closes #34042.

### Reason for this change

The action name in `Assets` stage has been changed from `FileAssetN` to the display name of the asset in #33844.

When pipeline's `publishAssetsInParallel` is false, the action name is also changed to the display name of the _first_ asset in stages instead of `FileAsset`.
The action uploads _all_ assets, therefore, its name should indicate "all assets" rather than the first asset.

### Description of changes

Use a specific (stable) display name when `singlePublisher` is true (pipeline's `publishAssetsInParallel` is false), instead of `stackAsset.displayName`.

### Describe any new or updated permissions being added

N/A

### Description of how you validated changes

Added a unit test and a integ test.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*