fix(core): splitting tree.json produces a lot of files (#34718)
In #34478, we made sure that `tree.json` would not exceed 512MB, by splitting parts of the construct tree to separate files.

The method we chose led to a lot of files.

In this PR we are combining the various small subtrees into larger "forest" files. Since there are no official readers for the new file format yet, we're not losing a lot of effort by changing the format.

The protocol is that if `treeId` is set in a subtree reference, the target file is a forest file, otherwise it's a tree file.

Closes #34627.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*