chore(s3): updated documentation to provide better insights when using grant methods for the `Bucket` resource (#34733)
### Issue # (if applicable)

Closes #34545 .

### Reason for this change

Current behavior in some `grant` methods for the `Bucket` resource might cause confusion and be seen as a bug.

### Description of changes

Added more details on the policy specifics, explaining why the current behavior is like that, and added additional resources if the implementer needs to restrict even more their permissions.


### Describe any new or updated permissions being added

N/A


### Description of how you validated changes

N/A. Just documentation being updated.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*