chore: add instructions to fix lfs errors (#34732)
### Reason for this change

A new Github action to check for LFS files was added and users without LFS cannot have their PR pass this test if they updated an integ test.

### Description of changes

Add instructions on how to check if git lfs is installed, how to install it and fix the problematic files.

### Description of how you validated changes

Helped a contributor make the test pass on his PR.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*