chore(pipelines): simplify CLI dependency version
The Pipelines construct used to inspect `package.json` of the library
to come up with the major version of the CDK CLI to use.

This made sense when we were dual-running v1 and v2, but now that v1 is
deprecated and the version lines of library and CLI have been decoupled,
this does not make sense anymore.

Just depend on CLI v2 directly, in a clear way.