feat(pipelines): cdk-assets version is configurable (#34802)
Make it possible to configure the version of `cdk-assets` CLI used by the CDK Pipeline.

This is only useful for versions of `cdk-assets` *newer*` than `3.3.1`, as older versions have wildcard ranges in their transitive dependencies and may install different versions of packages depending on what is currently available on npmjs.com.

The first `cdk-assets` version after `3.3.1` will bundle its dependencies so that the install operation will be deterministic.

We recommend you don't use this feature and just leave it at `latest`. The bundling of dependencies of `cdk-assets` will already remove any install failures around the release time of a new SDKv3 version, that used to plague `cdk-assets` in the past.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*