refactor(pipelines): simplify CLI dependency version (#34803)
The Pipelines construct used to inspect `package.json` of the library to come up with the major version of the CDK CLI to use.

This made sense when we were developing v1 and v2 of CDK at the same time, but now that v1 is deprecated and the version lines of library and CLI have been decoupled, this does not make sense anymore.

Just depend on CLI v2 directly, in a clear way.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*