chore: integ-runner uses toolkit-lib engine (#34809)
### Reason for this change

We are migrating `integ-runner` to use a new engine based of toolkit-lib instead of the CLI. To verify behavior, this engine was added under a new unstable feature flag. Let's use it in a repo that has a lot of traffic for `integ-runner` but is under own tight control and ownership, so we can easily react if something goes wrong.

### Description of changes

Use the new `toolkit-lib-engine` by delcaring `integ-runner --unstable=toolkit-lib-engine` on our yarn scripts.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

Tests still pass.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*