ci: cache docker images on main and use in PR builds (#34841)
### Reason for this change

> ECR Public repo allows 1 request/second for un-authenticated pulls and 10 request/second for authenticated pulls.

We try to avoid making requests to any repos by caching the used image as much as possible.
We do this on builds on main as a baseline and restore the cache on PRs. This assumes that most PRs won't add new images  and adding new images is rare enough.

### Description of changes

Add a cache for docker images.

### Describe any new or updated permissions being added

None.


### Description of how you validated changes

Run the commands manually.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*