fix(stepfunctions-tasks): properly serialize CallAwsServiceCrossRegion Lambda responses (#34843)
### Issue # (if applicable)

Closes #34768 

### Reason for this change

The handler in cross-region-aws-sdk-handler/index.ts file was not incorrectly returning byte-array as the payload when the called service was AWS Lambda. More details in the issue itself.

### Description of changes

The handler now checks the response to identify if it is a byte-array or other type of binary data and parses it correctly into a string, which is the expected format 

### Describe any new or updated permissions being added

NA


### Description of how you validated changes

Integ test added and unit tests

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*