fix(ecr-assets): incorrect handling of nested excludes in .dockerignore (#34810)
### Issue #13636

Closes #13636.

### Reason for this change

Fixes a bug where asset copying does not correctly process .dockerignore files. CDK did not copy files excluded by a nested exclude pattern in `.dockerignore`.

### Description of changes

Added a new method in `DockerIgnoreStrategy` that determines whether a directory is fully or partially ignored. The method is `DockerIgnoreStrategy.completelyIgnores` and if it returns true, it means that the directory is ignored AND that all of its children are ignored as well. For directories that are ignored but not "completely" ignored, the copyDirectory function will still walk through them.

As this changes the behaviour of asset copying, this can potientally be a breaking change. Therefore, a new feature flag is added which enables this fix. The feature flag is `DOCKER_IGNORE_NESTED_EXCLUDE_FIX`.

### Describe any new or updated permissions being added

No new permissions are added.

### Description of how you validated changes

Unit tests and integration tests are added.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*