feat(synthetics): add `maxRetries` prop to configure automatic retries on canary runs (#34541)
### Issue # (if applicable)

Closes #34511.

### Reason for this change
Adds support for configuring automatic retries of Canary runs.


### Description of changes
This adds a `maxRetries` property on the top-level of the CanaryProps interface. It uses this to configure the `retryConfig` structure on the `schedule` structure. This works similarly to how the `timeToLive` property configures the `durationInSeconds` field on the `schedule` structure.


### Describe any new or updated permissions being added
N/A



### Description of how you validated changes
I have updated the unit and integ tests, and have deployed manually, and verified the changes in the AWS console.


### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*