Converve LogGroupResourcePolicies in CloudWatchLogGroup
The `CloudWatchLogGroup` target is rather overzealous when creating a LogGroupResourcePolicy, especially given that CloudWatch has a hard limit of 10 per region. The way it is configured today, it is impossible to create more than 10 CloudWatch Log Group targets, as each creates its own LogGroupResourcePolicy.

This approach does two things. First, it makes the `resourcePolicyId` unique to the log group rather than the rule, since the LogGroupResourcePolicy only cares about the log group ARN, not the rule. Second, it tries to retrieve the policy from the stack node rather than the log group node. This second part is most likely a bug since it makes it impossible to find any existing policies since they're always created higher up in the construct hierarchy.